Only test Gear page if not testing all collection pages
I don't know if I am happy with this solution but it's the best I've got right now, and I have tried a LOT of different options. The same page is used for all tests when running in "limited" mode rather than each test using a random page, and the page used is always the same (Gear) rather than being random between runs. This reduces the potential test coverage of this spec a little but on the flip side will make any failing tests easier to debug. I don't know if this compromise is worth it though...