Ensure filters are in the correct state
The filters open expanded and automatically collapse as part of the page load. However, this can take a bit of time which can lead to flaky tests so add an assertion that the filter is collapsed before asserting further.

One could argue this should be a separate PR since it's not related to the test mode env var but I want to green up the pipeline rather than merging potentially flaky tests