Add test mode (#44)
* Add new env var

* Env var typing

* Extend collection page tests to use random/full test mode

This isn't working right - in random mode the test names can sometimes not be found in the worker process as they include a variable - but I want to get a commit in so that it's then easier to rework and revert to a known version if required

* Update env var definition

* Only test Gear page if not testing all collection pages

I don't know if I am happy with this solution but it's the best I've got right now, and I have tried a LOT of different options. The same page is used for all tests when running in "limited" mode rather than each test using a random page, and the page used is always the same (Gear) rather than being random between runs. This reduces the potential test coverage of this spec a little but on the flip side will make any failing tests easier to debug. I don't know if this compromise is worth it though...

* Update playwright.yml

* Correct describe title

The 'zzz' part of the name was added to help identify the tests when using a grep parameter on the CLI commands and shouldn't have been committed to the repo

* Parameterize product category test based on test mode env var

* Invert ternary logic

* Ensure filters are in the correct state

The filters open expanded and automatically collapse as part of the page load. However, this can take a bit of time which can lead to flaky tests so add an assertion that the filter is collapsed before asserting further.

One could argue this should be a separate PR since it's not related to the test mode env var but I want to green up the pipeline rather than merging potentially flaky tests