Rework CI pipelines (#61)
* Test on all browsers in parallel

* Remove shared setup job

I don't think there is an easy way of installing deps in one job and then running the tests in parallel in separate jobs, which is what I had initially planned. Let's start by trying to get the parallel test jobs running first and optimise later if possible

* Only install required browser per job

* Fix failing display options tests

* Skip sort test on Firefox & Webkit

* Reduce flake on Firefox further

* Increase timeout in page size test

This test is especially flaky on Firefox, presumably due to slow navigation on that browser for some reason. Hopefully by extending the assertion timeout we can reduce that flake. If this doesn't work then I will have to skip this test on Firefox

* Skip flaky test on Firefox

* Use matrix strategy for CI