Add test for similar products
I can't help but think there must be a better way to structure the test data for this e.g. integrate it with the product details since it is specific to a given product but that has proven beyond me at this stage due to circular references. Plus there is a chance that as I expand this testing to further products that the similar products will include some from different product categories which would complicate things even further if the similar products array were part of the individual product details