Product page tests pt1 (#66)
* Initial test

I don't think this model will be scalable across all products but at this stage I am just experimenting as how best to manage the test data and there will inevitably be some refactoring in the future

* Verify product details

* Extend product details test

* Add missing test data

* Extend product details test further

* Further test improvements

* Update tests for new data structure

* Tighten up types

* Add test for similar products

I can't help but think there must be a better way to structure the test data for this e.g. integrate it with the product details since it is specific to a given product but that has proven beyond me at this stage due to circular references. Plus there is a chance that as I expand this testing to further products that the similar products will include some from different product categories which would complicate things even further if the similar products array were part of the individual product details

* Add link tests

* Verify tab classes

* Refactor & extend product image testing

* Initial image carousel test

* Minor tweak to reduce flake

* Default thumbnail test

* Thumbnails change main image

* Cycle through product images

* Zoom in/out of product image

* Refactor bounding box verification

* Refactor img src verification

* Move test